+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: Add `color` prop for `TeamCardButtonWithTooltip` component.
+
 ## [1.1.0] - 2018-12-11
 
 Features: